Fix initial serverFrame position
Alberts Muktupāvels

I was not able to find bug for this on launchpad, but there is one for MATE:
https://github.com/mate-desktop/mate-panel/issues/249

This happens not only with mate-panel or gnome-panel, but also CSD windows. serverFrame is put at 0x0 position and must be clicked to be moved to correct position.

Same as https://code.launchpad.net/%7Emuktupavels/compiz/fix-initial-server-frame-position/+merge/322817